refactor(tests): add prefix to test suit names and reorganize tests types
